Port src/marks/{image,waffle} from JS to TS (stage 1)
Combine each .js and .d.ts pair into a single .ts file. Types/JSDoc come
from the .d.ts; runtime stays identical. Shared modules whose .d.ts files
omit runtime exports are imported with @ts-ignore per the established
Stage 1 convention.
